Edits

+++ botTesting/Currency/Stones.cs
@@ -10,8 +10,11 @@ using System.Threading.Tasks;
 
 namespace botTesting.Currency
 {
+
     public class Stones : ModuleBase<SocketCommandContext>
     {
+        public static List<DateTimeOffset> stackCooldownTimer = new List<DateTimeOffset>();
+        public static List<SocketGuildUser> stackCooldownTarget = new List<SocketGuildUser>();
         [Group("money")]
         public class StonesGroup : ModuleBase<SocketCommandContext>
         {
@@ -125,32 +128,37 @@ namespace botTesting.Currency
             [Command("reset")]
             public async Task Reset(IUser User = null)
             {
-                 if(User == null)
+                SocketGuildUser ChkUser = Context.User as SocketGuildUser;
+                if (ChkUser.GuildPermissions.Administrator)
                 {
-                    await Context.Channel.SendMessageAsync("Specify a user to reset bruh");
-                    return;
-                }
-                if (User.IsBot)
-                {
-                    await Context.Channel.SendMessageAsync("Stop including bots :neutral_face:");
-                    return;
-                }
-
-                using(var DbContext = new SQLiteDBContext())
-                {
-                    int Num= DbContext.Stones.Where(x => x.UserId == User.Id).Count();
-                    if (Num == 0)
+                    if (User == null)
                     {
-                        await Context.Channel.SendMessageAsync($"{User.Mention} was already deleted");
+                        await Context.Channel.SendMessageAsync("Specify a user to reset bruh");
                         return;
                     }
-                    else
+                    if (User.IsBot)
                     {
-                        await Context.Channel.SendMessageAsync("You got reset, so you lost all your money :cry:");
-                        DbContext.Stones.RemoveRange(DbContext.Stones.Where(x => x.UserId == User.Id));
-                        await DbContext.SaveChangesAsync();
+                        await Context.Channel.SendMessageAsync("Stop including bots :neutral_face:");
+                        return;
+                    }
+
+                    using (var DbContext = new SQLiteDBContext())
+                    {
+                        int Num = DbContext.Stones.Where(x => x.UserId == User.Id).Count();
+                        if (Num == 0)
+                        {
+                            await Context.Channel.SendMessageAsync($"{User.Mention} was already deleted");
+                            return;
+                        }
+                        else
+                        {
+                            await Context.Channel.SendMessageAsync("You got reset, so you lost all your money :cry:");
+                            DbContext.Stones.RemoveRange(DbContext.Stones.Where(x => x.UserId == User.Id));
+                            await DbContext.SaveChangesAsync();
+                        }
                     }
                 }
+                await Context.Channel.SendMessageAsync("Not a mod retard");
             }
             [Command("buyDogs")]
             public async Task BuyDogs()
@@ -174,20 +182,32 @@ namespace botTesting.Currency
             }
             [Command("work")]
             public async Task Work()
-            {                  
-                using (var DbContext = new SQLiteDBContext()) {
-                    Random Rand = new Random();
-                    string[] Jobs = {"You worked at a factory", "You worked at a hotel",
-                                "You worked as a chef", "You worked at a graveyard"}; //add more jobs
-                    int Job = Rand.Next(Jobs.Length);
-                    int Cash = Rand.Next(201) + 300;
-                    Stone GiveCash = DbContext.Stones.Where(x => x.UserId == Context.User.Id).FirstOrDefault();
-                    GiveCash.Amount += Cash;
-                    await Context.Channel.SendMessageAsync(Jobs[Job]);    
-                    DbContext.Update(GiveCash);
-                    await DbContext.SaveChangesAsync();
-                    await Context.Channel.SendMessageAsync("You earned $" + Cash);
-                }   
+            {
+                if (stackCooldownTarget.Contains(Context.User as SocketGuildUser))
+                {
+                    if (stackCooldownTimer[stackCooldownTarget.IndexOf(Context.Message.Author as SocketGuildUser)].AddSeconds(30) >= DateTimeOffset.Now)
+                    {
+                        int secondsLeft = (int)(stackCooldownTimer[stackCooldownTarget.IndexOf(Context.Message.Author as SocketGuildUser)].AddSeconds(30) - DateTimeOffset.Now).TotalSeconds;
+                        EmbedBuilder Embed = new EmbedBuilder();
+                        Embed.WithAuthor(Context.User.Username, Context.User.GetAvatarUrl());
+                        Embed.WithColor(40, 200, 150);
+                        Embed.WithDescription($"Wait {secondsLeft} more seconds before working dumbass");
+                        await Context.Channel.SendMessageAsync("", false, Embed.Build());
+                        return;
+                    }
+                    else
+                    {
+                        stackCooldownTimer[stackCooldownTarget.IndexOf(Context.Message.Author as SocketGuildUser)] = DateTimeOffset.Now;
+                        await WorkMethod();
+                    }
+                }
+                else
+                {
+                    stackCooldownTarget.Add(Context.User as SocketGuildUser);
+                    stackCooldownTimer.Add(DateTimeOffset.Now);
+                    await WorkMethod();
+                }
+                
             }
             [Command("inventory")]
             public async Task Inventory()
@@ -203,6 +223,26 @@ namespace botTesting.Currency
                         Embed.AddField("Dogs:", Inv.Item1);
                         await Context.Channel.SendMessageAsync("", false, Embed.Build());
                     }
+                }
+            }
+            public async Task WorkMethod()
+            {
+                using (var DbContext = new SQLiteDBContext())
+                {
+
+                    Random Rand = new Random();
+                    string[] Jobs = {"You worked at a factory", "You worked at a hotel",
+                                "You worked as a chef", "You worked at a graveyard",
+                                "You did chores because no women were there to help", "You got away **just** in time from a bank robbery",
+                                "Before the owners could find out, you robbed their mansion", "You became a sex slave for a day, pleasing several people",
+                                "You were on your way to work but got hit by a car but get compensated", "You get hired to murder"}; //add more jobs
+                    int Job = Rand.Next(Jobs.Length);
+                    int Cash = Rand.Next(401) + 100;
+                    Stone GiveCash = DbContext.Stones.Where(x => x.UserId == Context.User.Id).FirstOrDefault();
+                    GiveCash.Amount += Cash;
+                    await Context.Channel.SendMessageAsync(Jobs[Job] + ", earning $" + Cash);
+                    DbContext.Update(GiveCash);
+                    await DbContext.SaveChangesAsync();
                 }
             }
